requirements.txt
openai>=0.27.0
python-dotenv>=0.21.0
pytest>=7.0.0
# Optional: to create a single-file executable using PyInstaller
# pyinstaller>=5.0
cryptography>=41.0.0
pyperclip>=1.8.2
pystray>=0.19.5
Pillow>=10.0.0
requests>=2.30.0
beautifulsoup4>=4.12.2
